requirements.txt
google-cloud-aiplatform
google-cloud-documentai
functions-framework
chromadb
gcsfs
functions-framework
google-cloud-storage
langchain-google-vertexai
